Fix calibration_factor precision loss; add explicit per-dose re-tare
calibration_factor was stored/transmitted as float, but this
hardware's real calibration factor (~1e-4 magnitude, ~6-7 significant
digits) doesn't fit float32's ~7 total significant digits once
combined with a multiplication against raw ADC deltas in the tens of
thousands -- and ArduinoJson's double serializer further truncates
values this small to ~6 significant digits over the wire regardless of
storage type. Switches calibration_factor to double end-to-end
(ADS1232::calFactor, SettingsSnapshot, NVS, the WS wire format) and
scales it by 1e6 on the wire (calibration_factor_x1e6) so the same
fixed-decimal-places JSON serialization has real digits to work with.

Also adds an explicit hardware re-tare request (Dosing -> Scale) fired
the instant a dose is confirmed, so every dose starts from a freshly-
zeroed ADC baseline instead of whatever the idle auto-tare last left
behind. A new debug-only TARE_DEBUG telemetry event and
v2.tare_debug table log each session's tare baseline (raw ADC count +
grams) so the same physical dosing cup's tare consistency can actually
be checked against real data. FINALIZE now also detects a large weight
swing (the cup being lifted) and dismisses straight to IDLE rather than
letting the fixed-width delta readout overflow.

A new Scale -> Network latest-sample mailbox plus a raw_read_request/
raw_read WS message pair support live on-demand ADC polling for an
upcoming Advanced/Calibration debug page.

+++ lib/ADS1232/ADS1232.cpp
@@ -28,7 +28,7 @@ ADS1232::ADS1232(uint8_t pdwn, uint8_t sclk, uint8_t dout, uint8_t spd,
       spdPin(spd),
       gain1Pin(gain1pin),
       gain0Pin(gain0pin),
-      calFactor(1.0f),
+      calFactor(1.0),
       ringBufferIndex(0),
       ringBufferSize(1) {
   resetBuffer();
@@ -184,12 +184,15 @@ bool ADS1232::tare() {
   return isStable;
 }
 
-void ADS1232::setCalFactor(float cal) { calFactor = cal; }
+void ADS1232::setCalFactor(double cal) { calFactor = cal; }
 
 double ADS1232::getUnits() {
   bool isStable = false;
-  float raw = getRaw(isStable) - tareRaw;
-  float units = raw * calFactor;
+  // double throughout, not just on the return type -- computing the
+  // multiplication itself in float would throw away calFactor's extra
+  // precision right here regardless of what type carried it in.
+  double raw = getRaw(isStable) - tareRaw;
+  double units = raw * calFactor;
 
   // First run (or invalid last) always accept
   if (isnan(_lastUnitsFiltered)) {
